fix(calendar): use Long instead of Int for calendar/event IDs on Android
Fixes NumberFormatException when calendar/event/attendee IDs exceed
Integer.MAX_VALUE (2,147,483,647). This can happen on some Android
devices with synced Google calendars where IDs can be very large.

Changed all ID parsing from `.toInt()` to `.toLong()` and updated
function return types from `Int` to `Long` where applicable.

This is backward compatible since all IDs are converted to strings
via `.toString()` before being sent to JavaScript.

Fixes #43343

+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/CalendarModule.kt
@@ -217,7 +217,7 @@ class CalendarModule : Module() {
 
   //endregion
 
-  private suspend fun saveEvent(eventInput: Either<NewEventInput, EventUpdateInput>): Int {
+  private suspend fun saveEvent(eventInput: Either<NewEventInput, EventUpdateInput>): Long {
     return if (eventInput.`is`(EventUpdateInput::class)) {
       eventRepository.updateEvent(eventInput.second())
     } else {
